Proj/blogv2/requirement.txt
setuptools==45.2.0
aliyun-python-sdk-core==2.13.5
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-kms==2.7.1
amqp==2.5.2
astroid==2.2.5
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
decorator==4.4.1
Django==2.2.3
django-cors-middleware==1.4.0
django-debug-toolbar==2.0
django-filter==2.1.0
django-guardian==2.0.0
django-qiniu-storage==2.3.1
djangorestframework==3.10.0
geoip2==2.9.0
idna==2.8
importlib-metadata==1.1.0
ipython==7.10.1
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.6.6
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
maxminddb==1.4.1
mccabe==0.6.1
more-itertools==8.0.0
mysqlclient==1.4.2.post1
oss2==2.8.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
ptyprocess==0.6.0
pycryptodome==3.8.2
Pygments==2.4.2
pylint==2.3.1
pytz==2019.1
qiniu==7.2.6
redis==3.3.11
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.2
zipp==0.6.0
